% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cups2TbspFunction.R
\name{cups2Tbsp}
\alias{cups2Tbsp}
\title{Cups to tablespoons conversion}
\usage{
cups2Tbsp(x)
}
\arguments{
\item{x}{A numeric value representing the number of cups needed}
}
\value{
The number of tablespoons that are equal to the number of cups that were inputted into the function
}
\description{
Take a numeric input as the number of cups and converts it to tablespoons
}
