% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gals2cupsFunction.R
\name{gals2cups}
\alias{gals2cups}
\title{Gallons to cups conversion}
\usage{
gals2cups(x)
}
\arguments{
\item{x}{A numeric value representing the number of gallons needed}
}
\value{
The number of cups that are equal to the number of gallons that were inputted into the function
}
\description{
Takes a numeric input as the number of gallons and converts it to cups
}
