% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gals2pintsFunction.R
\name{gals2pints}
\alias{gals2pints}
\title{Gallons to pints conversion}
\usage{
gals2pints(x)
}
\arguments{
\item{x}{A numeric value representing the number of gallons needed}
}
\value{
The number of pints that are equal to the number of gallons that were inputted into the function
}
\description{
Takes a numeric input as the number of gallons and converts it to pints
}
