% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gals2quartsFunction.R
\name{gals2quarts}
\alias{gals2quarts}
\title{Gallons to quarts conversion}
\usage{
gals2quarts(x)
}
\arguments{
\item{x}{A numeric value representing the number of gallons needed}
}
\value{
The number of quarts that are equal to the number of gallons that were inputted into the function
}
\description{
Takes a numeric input as the number of gallons and converts it to quarts
}
