% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pints2cupsFunction.R
\name{pint2cups}
\alias{pint2cups}
\title{Pints to cups conversion}
\usage{
pint2cups(x)
}
\arguments{
\item{x}{A numeric value representing the number of pints needed}
}
\value{
The number of cups that are equal to the number of pints that were inputted into the function
}
\description{
Takes a numeric input as the number of pints and converts it to cups
}
