% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarts2pintsFunction.R
\name{quart2pints}
\alias{quart2pints}
\title{Quarts to pints conversion}
\usage{
quart2pints(x)
}
\arguments{
\item{x}{A numeric value representing the number of quarts needed}
}
\value{
The number of pints that are equal to the number of quarts that were inputted into the function
}
\description{
Takes a numeric input as the number of quarts and converts it to pints
}
