% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarts2cupsFunction.R
\name{quarts2cups}
\alias{quarts2cups}
\title{Quarts to cups conversion}
\usage{
quarts2cups(x)
}
\arguments{
\item{x}{A numeric value representing the number of quarts needed}
}
\value{
The number of cups that are equal to the number of quarts that were inputted into the function
}
\description{
Takes a numeric input as the number of quarts and converts it to cups
}
