% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tbsp2tspFunction.R
\name{Tbsp2tsp}
\alias{Tbsp2tsp}
\title{Tablespoons to teaspoons conversion}
\usage{
Tbsp2tsp(x)
}
\arguments{
\item{x}{A numeric value representing the number of Tablespoons needed}
}
\value{
The number of teaspoons that are equal to the number of Tablespoons that were inputted into the function
}
\description{
Takes a numeric input as the number of Tablespoons and converts it to teaspoons
}
